split 'diff' and 'patch' (now called 'act')
I have realized that starting from the 'diff' method was a mistake.

The main reason is that I want to be able to call 'patch' on types which
contain callbacks, but I don't want the type system to claim that
diffing IO actions is possible.

Therefore, I need to split 'diff' and 'patch' into two separate
typeclasses. Should those be two independent 'ToPatch' and 'FromPatch'
typeclasses, or should one be a specialization of the other? Where would
the associated 'Patch' type and its 'Monoid' constraint go?

I then remembered about the mathematical concept of a "group action",
and was happy to discover that a "monoid action" with the laws I want
for 'patch' is a well-defined mathematical concept as well. So it would
probably make more sense if there was an 'Action' typeclass specializing
'Monoid', and a 'Diff' typeclass specializing 'Action'.