WindowRef, ButtonRef
It's an FLTK.Ref FLTK.Button under the hood, but I named it Button to
hide the implementation details. Now I need a name for the state of the
button, and I'd rather use Button than ButtonState, so I'm renaming
Button to ButtonRef in order to make the Button name available.